Geofacing - 1

+++ app/src/main/java/pl/devone/android/mapboxexampleapp/models/PastLocation.kt
@@ -6,4 +6,4 @@ import com.mapbox.mapboxsdk.geometry.LatLng
 /**
  * Created by ljedrzynski on 22.01.2018.
  */
-data class PastLocation(val location: Location, val name: String, val description: String, val radius: Int) : LatLng(location.latitude, location.longitude)
+data class PastLocation(val location: Location, val name: String, val description: String, val radius: Float) : LatLng(location.latitude, location.longitude)
